backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0
requests>=2.31.0

# Image Processing
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
Pillow>=10.0.0
numpy>=1.21.0,<2.0.0

# ML/AI
mediapipe>=0.10.0
rembg>=2.0.50
onnxruntime

# Database
sqlalchemy>=2.0.0

# Environment
python-dotenv>=1.0.0
setuptools<70.0.0
llvmlite==0.42.0
numba==0.59.1

# Diffusion dependencies
torch
torchvision
diffusers
transformers
accelerate
scipy
